Modified certain datatypes
Some getters and setters had int data type when the original object was a double. Those were fixed.

+++ Backend/src/main/java/com/TermProject/Banking/model/account.java
@@ -2,6 +2,8 @@ package com.TermProject.Banking.model;
 
 import jakarta.persistence.*;
 
+import java.util.Optional;
+
 @Entity
 public class account {
     @Id
@@ -29,7 +31,7 @@ public class account {
         return balance;
     }
 
-    public void setBalance(int balance) {
+    public void setBalance(double balance) {
         this.balance = balance;
     }
